Server/requirements.txt
Flask
Flask-PyMongo
Flask-CORS
python-dotenv
Flask-JWT-Extended
pymongo
flask_socketio
ultralytics
pathlib